+++ app/src/operations/exec/index.test.ts
@@ -1,22 +1,6 @@
-import { createTestingPinia } from '@pinia/testing';
-import { setActivePinia } from 'pinia';
-import { beforeEach, describe, expect, it, vi } from 'vitest';
-
-import { cryptoStub } from '@/__utils__/crypto';
-
-vi.stubGlobal('crypto', cryptoStub);
-
-import { useServerStore } from '@/stores/server';
+import { describe, expect, it } from 'vitest';
 
 import config from './index';
-
-beforeEach(() => {
-	setActivePinia(
-		createTestingPinia({
-			createSpy: vi.fn,
-		})
-	);
-});
 
 describe('Overview', () => {
 	it('Renders empty array', () => {
@@ -25,17 +9,7 @@ describe('Overview', () => {
 });
 
 describe('Options', () => {
-	it("Doesn't show notice when no modules are allowed", () => {
+	it('Returns the code field as the only option', () => {
 		expect(config.options()).toHaveLength(1);
-	});
-
-	it('Shows notice when modules are allowed', () => {
-		const serverStore = useServerStore();
-
-		serverStore.info.flows = {
-			execAllowedModules: ['nanoid'],
-		};
-
-		expect(config.options()).toHaveLength(2);
 	});
 });